Moved the hover popop event to the other events
